# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Janora API",
    "description": "Residential Access Management Platform",
    "version": "1.0",
    "contact": {
      "name": "Janora Team",
      "email": "support@janora.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateVisitorDto": {
        "type": "object",
        "required": ["firstName", "lastName", "documentType", "documentNumber", "propertyId"],
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "documentType": {
            "type": "string",
            "enum": ["CC", "CE", "PASSPORT", "NIT", "TI"]
          },
          "documentNumber": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "propertyId": {
            "type": "string"
          }
        }
      },
      "Visitor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "documentType": {
            "type": "string"
          },
          "documentNumber": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "propertyId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/v1/visitors": {
      "get": {
        "tags": ["Visitors"],
        "summary": "Listar visitantes",
        "operationId": "findAllVisitors",
        "parameters": [
          {
            "name": "propertyId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de visitantes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Visitor"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Visitors"],
        "summary": "Registrar visitante",
        "operationId": "createVisitor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVisitorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Visitante creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Visitor"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/visitors/{id}": {
      "get": {
        "tags": ["Visitors"],
        "summary": "Obtener visitante por ID",
        "operationId": "findOneVisitor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Visitante encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Visitor"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Visitors"],
        "summary": "Actualizar visitante",
        "operationId": "updateVisitor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVisitorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Visitante actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Visitor"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Visitors"],
        "summary": "Eliminar visitante (soft delete)",
        "operationId": "removeVisitor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Visitante eliminado"
          }
        }
      }
    },
...
  }
}